pyproject.toml
[tool.poetry]
name = "e-net"
version = "0.1.0"
description = ""
authors = ["adieulatete"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
djangorestframework = "^3.15.2"
django = "^5.1.1"
psycopg2-binary = "^2.9.9"
djangorestframework-simplejwt = "^5.3.1"
celery = "^5.4.0"
redis = "^5.0.8"
qrcode = "^7.4.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
